created new methods to decrease repeated code
