+++ frontend/src/components/navbar/navbar.js
@@ -4,7 +4,7 @@ import './navbar.css'
 import LoggedOutBar from './navbar_content/loggedOut';
 import LoggedInBar from './navbar_content/loggedIn';
 import SpotifyWebApi from 'spotify-web-api-js';
-import { checkIfRefreshTokenWorking, checkIfTokenActive } from '../../scripts/checkTokens';
+import { checkIfRefreshTokenWorking, checkIfTokenActive, generateNewAccessToken, removeTokens } from '../../scripts/handleTokens';
 import { decode, encode } from '../../scripts/encoder';
 import generateRandomString from '../../scripts/randomString';
 
@@ -24,7 +24,7 @@ class NavBar extends React.Component {
         let refresh_token = localStorage.getItem('refresh_token');
         let access_token = decode(localStorage.getItem('access_token'), key);
 
-        if(refresh_token) {
+        if(refresh_token) { // if we have refresh token
             checkIfRefreshTokenWorking(refresh_token)
             .then(res => {
                 this.setState({
@@ -34,21 +34,12 @@ class NavBar extends React.Component {
             .catch(err => console.log(err))
 
             checkIfTokenActive(access_token, spotifyApi)
-            .catch(err => {
-                fetch('http://localhost:8888/refresh_token?refresh_token='+refresh_token)
-                .then(response => response.json())
-                .then(data => {
-                    key = generateRandomString(15);
-                    localStorage.setItem('key', key);
-                    localStorage.setItem('access_token', encode(data.access_token, key));
-                })
+            .catch(err => { // if access token not active generate new one
+                generateNewAccessToken(refresh_token)
             })
         }
-
-        else {
-            localStorage.removeItem('refresh_token');
-            localStorage.removeItem('access_token');
-            localStorage.removeItem('key');
+        else { // if we don't have refresh token
+            removeTokens() // login again
         }
     }
